PR feedback

+++ components/Snackbar/tests/unit/MDCSnackbarMessageViewTests.m
@@ -254,7 +254,7 @@
 
 - (void)testManagerForwardsButtonProperties {
   // Given
-  self.manager.disabledButtonAlpha = 0.5f;
+  self.manager.disabledButtonAlpha = (CGFloat)0.5;
   self.manager.uppercaseButtonTitle = NO;
   self.manager.buttonInkColor = UIColor.redColor;
   MDCSnackbarMessageAction *action = [[MDCSnackbarMessageAction alloc] init];
